+++ src/sdk/PnP.Core/Model/SharePoint/Core/Internal/UserCustomAction.cs
@@ -10,6 +10,7 @@ namespace PnP.Core.Model.SharePoint
     /// </summary>
     [SharePointType("SP.UserCustomAction", Target = typeof(Web), Uri = "_api/Web/UserCustomActions('{Id}')", Get = "_api/Web/UserCustomActions", LinqGet = "_api/Web/UserCustomActions")]
     [SharePointType("SP.UserCustomAction", Target = typeof(Site), Uri = "_api/Site/UserCustomActions('{Id}')", Get = "_api/Site/UserCustomActions", LinqGet = "_api/Site/UserCustomActions")]
+    [SharePointType("SP.UserCustomAction", Target = typeof(List), Uri = "_api/Web/Lists(guid'{Parent.Id}')/UserCustomActions('{Id}')", Get = "_api/Web/Lists(guid'{Parent.Id}')/UserCustomActions", LinqGet = "_api/Web/Lists(guid'{Parent.Id}')/UserCustomActions")]
     internal sealed class UserCustomAction : BaseDataModel<IUserCustomAction>, IUserCustomAction
     {
         internal const string AddUserCustomActionOptionsAdditionalInformationKey = "AddOptions";