+++ src/sdk/CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Improve code consistency for the eventhub model [jansenbe - Bert Jansen]
 - Added `DefaultViewUrl` property for `IList` [jansenbe - Bert Jansen]
 - Added `PublishingPagesLibrary` as possible `ListTemplateType` [jansenbe - Bert Jansen]
+- Added `UserCustomActions` to `IList` [jansenbe - Bert Jansen]
 
 ## [1.8.0]